modified event and member collection/adapter/model so its usable globally
