+++ Android App/ColonelFund/app/src/main/java/com/colonelfund/colonelfund/MemberCollection.java
@@ -272,4 +272,22 @@ public class MemberCollection {
     public Collection<Member> getMembersList() {
         return memberMap.values();
     }
+
+    /**
+     * Generates Initials and User Name for member list.
+     *
+     * @return array list of models for member list
+     */
+    public ArrayList<MemberListModel> generateListData() {
+        ArrayList<MemberListModel> models = new ArrayList<MemberListModel>();
+        Iterator<Member> memberIter = memberMap.values().iterator();
+        while (memberIter.hasNext()) {
+            Member temp = memberIter.next();
+            String firstName = temp.getFirstName();
+            String lastName = temp.getLastName();
+            models.add(new MemberListModel(firstName.substring(0, 1) + lastName.substring(0, 1),
+                    temp.getUserID(), firstName, lastName, temp.getEmailAddress()));
+        }
+        return models;
+    }
 }
